Fix compile issue and warning
1. remove unused head files which cause compile fail
2. Remove compile warning

+++ FreeRTOS/Demo/T-HEAD_CB2201_CDK/RTOSDemo_CDK/RTOSDemo/main.c
@@ -82,7 +82,7 @@ static void prvQueueSendTask( void *pvParameters )
 
 static void prvQueueReceiveTask( void *pvParameters )
 {
-	unsigned long ulReceivedValue;
+	unsigned int ulReceivedValue;
 
 	for( ;; )
 	{
